backend/Financogram/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo[srv]==3.12.3
yfinance==0.2.65
numpy==1.24.3
pandas==1.5.3
requests==2.31.0
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
mediapipe==0.10.11
deepface==0.0.83
tensorflow==2.13.0
tf_keras==2.19.0
keras==2.13.1
opencv-python==4.8.0.76
beautifulsoup4==4.13.4
pillow==11.3.0
imageio==2.37.0
qrcode==8.2
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
openpyxl
python-decouple
django-environ
gunicorn

